Renamed module for Exercise 8-15 to a more apropriate naming convention according to PEP8 guidelines
