+++ Ch. 8 Exercises/Exercise_8_15 - printing_models.py
@@ -1,7 +1,7 @@
 # Add the functions for the example code 'printing_models.py' into a separate 
 # file called 'printing_functions.py'. Add an import statement to the top of 
 # 'printing_models.py' and call the functions.
-import 8_15_printing_functions
+import printing_functions
 
 unprinted_designs = ['phone case', 'robot pendant', 'dodecahedron']
 completed_models = []
