+++ Ch. 8 Exercises/printing_functions.py
